Add support for memory-related bulk-memory instructions
This commit adds support for some new instructions:

* `data.drop`
* `memory.init`
* `memory.copy`
* `memory.fill`

This also adds support for passive data and the data count section. A
new `Config::bulk_memory_enabled` boolean, defaulting to `false`, is
added which gates all of this support. Finally the generation of data
segments was moved before the code so we know whether `data.drop` will
be valid by the time we're generating code.

Finally, a new fuzz target was added which validates all possible
modules (generated by `SwarmConfig`) with `wasmparser`.